Added improvements
